refactor: add arial-labels for accessibility
