+++ src/components/navbar/Navbar.tsx
@@ -11,18 +11,18 @@ import {
   NavigationMenuTrigger,
   navigationMenuTriggerStyle,
 } from "@/components/ui/navigation-menu";
-import logo from "@/assets/KeySystem.svg";
-import phoneIcon from "@/assets/icons/phone.svg";
+import logo from "@/assets/KeySystem.webp";
+import phoneIcon from "@/assets/icons/phone.webp";
 import MobileNavbar from "./MobileNavbar";
 
 const Navbar = () => {
   return (
     <header className="sticky top-0 left-0 right-0 z-50 bg-gradient-to-r from-[#05071E] to-[#000080] text-[22px] font-normal shadow-sm text-[#CFCFCF]">
-      <div className="container mx-auto flex items-center justify-between  py-[10px]">
+      <div className="container mx-auto flex items-center justify-between  py-[10px]" aria-label="Main navigation">
         {/* Logo */}
-        <div className="flex items-center gap-2">
+        <a className="flex items-center gap-2" aria-label="Go to homepage">
           <img src={logo} alt="KeySystem Logo" className="h-[50.04px]" />
-        </div>
+        </a>
 
         {/* Desktop Navigation */}
         <NavigationMenu>
@@ -31,7 +31,7 @@ const Navbar = () => {
               <NavigationMenuItem key={link.name}>
                 {link.children ? (
                   <>
-                    <NavigationMenuTrigger className="bg-transparent font-normal">
+                    <NavigationMenuTrigger className="bg-transparent font-normal" aria-haspopup="true" aria-expanded="false">
                       {link.name}
                     </NavigationMenuTrigger>
                     <NavigationMenuContent>
@@ -71,7 +71,7 @@ const Navbar = () => {
         {/* Call Button */}
         <a
           href="tel:+2348184441404"
-          className="hidden md:flex justify-center items-center gap-[6.78px] rounded-full bg-gradient-to-r from-[#000080] via-[#4169E1] to-[#000080] px-[27.1px] py-[12.2px] border-[1.36px] border-[#4169E1] text-[16.26px] font-bold text-white shadow hover:bg-gradient-to-r transition"
+          className="hidden md:flex justify-center items-center gap-[6.78px] rounded-full bg-gradient-to-r from-[#000080] via-[#4169E1] to-[#000080] px-[27.1px] py-[12.2px] border-[1.36px] border-[#4169E1] text-[16.26px] font-bold text-white shadow hover:bg-gradient-to-r transition" aria-label="Call KeySystem on +234 818 444 1404"
         >
           <img src={phoneIcon} alt="phone icon" className="w-4 h-4" />
           <span>+234 818 444 1404</span>
